Helper for parsing simple response error
Adds error message to redux object
