+++ src/components/Layout/index.js
@@ -1,7 +1,9 @@
 import Column from './Column'
 import Row from './Row'
+import Base from './Base'
 
 export {
   Column,
-  Row
+  Row,
+  Base
 }
